Add extraction stamps, fix BIGINT model drift, close index drift
- audience_interactions gains extracted_at/extracted_prompt_version/
  extracted_model so batch extraction marks every comment in a chunk
  as done, not just the anchor — re-runs no longer re-extract the
  same comments (duplicate inferred observations, wasted LLM calls).
  Extraction pipeline pre-flight now checks the stamp first and
  back-fills stamps for rows already covered by evidence.
- creator_platform_accounts.total_view_count: model said Integer but
  the DB column is BIGINT; large channels exceed int32's 2.1B cap.
  Model now matches (BigInteger).
- CommercialSignal model declares ix_signals_cluster_id and
  ix_signals_evidence_id, matching the indexes the DB already has,
  so autogenerate stops proposing to drop them.
- New migration 6c88964d1240 adds only the three stamp columns;
  up/down/up verified, autogenerate is drift-free afterwards.
- test_decision_validation rewritten for the URL-scoped decision
  contract (404 unknown creator, 422 bad enum, extra fields ignored).

Full suite: 748 passed, 1 skipped.

+++ corp/core/models/creator.py
@@ -2,7 +2,7 @@ import enum
 from datetime import datetime
 from typing import TYPE_CHECKING
 
-from sqlalchemy import DateTime, Enum, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, Enum, ForeignKey, Index, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from corp.core.models.base import Base, TimestampMixin, generate_uuid
@@ -73,7 +73,8 @@ class CreatorPlatformAccount(TimestampMixin, Base):
     # Point-in-time channel enrichment. MetricsSnapshot (metrics.py) tracks these
     # over time for growth/velocity scoring; these columns hold the latest value
     # for cheap reads that don't need history.
-    total_view_count: Mapped[int | None] = mapped_column(Integer)
+    # BigInteger: total view counts on large channels exceed int32's 2.1B cap.
+    total_view_count: Mapped[int | None] = mapped_column(BigInteger)
     video_count: Mapped[int | None] = mapped_column(Integer)
     joined_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     country: Mapped[str | None] = mapped_column(String(10))
